Маленькое обновление
Заменил уфатоса на себя в коде.
Добавил библиотеки
Добавил данные проекта
И вообще проверяю как работает netBeans со всем этим
