+++ RPChat/lib/nblibraries.properties
@@ -1,0 +1,2 @@
+libs.CopyLibs.classpath=\
+    ${base}/CopyLibs/org-netbeans-modules-java-j2seproject-copylibstask.jar
