+++ RPChat/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=ba0b4cda
+build.xml.script.CRC32=2bd5114f
+build.xml.stylesheet.CRC32=28e38971@1.44.1.45
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=ba0b4cda
+nbproject/build-impl.xml.script.CRC32=769dbd3d
+nbproject/build-impl.xml.stylesheet.CRC32=0ae3a408@1.44.1.45
