--- src/ru/ufatos/chat/config.yml
@@ -1,31 +1,0 @@
-Range:
-   main: 250
-   whispering: 20
-   shout: 500
-disableDeathMessage: true
-Id:
-   worldId: 260
-   worldSubId: 0
-   globalId: 264
-   globalSubId: 0
-#------------------------------------------
-#Range:
-#   main: 500 - локальный чат
-#   whispering: 15 - шёпот
-#   shout: 1000 - крик
-#disableDeathMessage: true - скрыть ли сообщения о смерти
-#Id:
-#   worldId: 260 - в мир
-#   worldSubId: 0 - в мир
-#   globalId: 264 - в глобал
-#   globalSubId: 0 - в глобал
-#
-# Permission
-#	mcnw.global
-#	mcnw.global.no_item
-#	mcnw.world
-#	mcnw.world.no_item
-#	mcnw.nospy
-#	mcnw.spy
-#	mcnw.pmspy
-#------------------------------------------
